chore: add loadEnv test & refactor dev-server
