+++ packages/server/core/tests/base/fixtures/serverEnv/.env
@@ -1,0 +1,1 @@
+USER_NAME=root
