+++ packages/server/core/tests/base/fixtures/serverEnv/.env.prod
@@ -1,0 +1,2 @@
+USER_NAME=prod_root
+ENV=prod
