+++ packages/server/core/tests/base/fixtures/serverEnv/.env.test
@@ -1,0 +1,2 @@
+USER_NAME=test_root
+ENV=test
